Added CRUD operations for Admin Categories and Users

+++ resources/views/categories/indexC.blade.php
@@ -10,24 +10,24 @@
             <th>Created On</th>
             <th colspan="2">Action</th>
         </tr>
-        <tr>
+        
         @if(count($categories))
             @foreach($categories as $category)
+            <tr>
                 <td>{{ $category->id }}</td>
                 <td>{{ $category->category_parent }}</td>
                 <td>{{ $category->category_name }}</td>
                 <td>{{ $category->created_at->toFormattedDateString() }}</td>
                 <td>
-                    <a href="/categories/edit" class="btn btn-warning">Edit</a>
+                    <a href="/categories/edit/{{ $category->id }}" class="btn btn-warning">Edit</a>
                 </td>
                 <td>
-                    <a href="/categories/delete" class="btn btn-danger">Delete</a>
+                    <a href="/categories/delete/{{ $category->id }}" class="btn btn-danger" onclick="return confirm('Are you sure you want to delete this category?')">Delete</a>
                 </td>
-
+            </tr>
             @endforeach
-        @endif
-            
-        </tr>
+        @endif       
+       
     
     </table>
